refactor(localFileUtils): move out form instructions folder
